pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hms-silver-solution"
version = "0.1.0"
description = "Reproducible package for a Kaggle Silver Medal HMS harmful brain activity classification solution."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Yuto Terashima", email = "yutoterashima4@gmail.com" }]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "matplotlib>=3.7",
  "pyyaml>=6.0",
  "nbformat>=5.9",
  "scipy>=1.10",
]

[project.optional-dependencies]
gpu = [
  "torch>=2.1",
  "timm>=0.9",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

